pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lightspeed-to-dataverse-exporter"
version = "0.1.0"
description = "Export Lightspeed data to Dataverse ingress service"
authors = [
    {name = "Ondrej Metelka", email = "ometelka@redhat.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0.0",
    "requests>=2.25.0",
    "pyyaml>=6.0.0",
    "kubernetes>=33.1.0",
]

[project.scripts]
lightspeed-exporter = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
] 
